Rename the plugin to cloud-storage-archive
